updated Overview of Title Storage
Changed the list of platforms that can read/write to Universal Platform storage--Xbox One consoles can read from Universal (that was a bug in the docs), and there is no point in mentioning X360 or WinPhone because neither of those platforms are seeing active development. This way it reads much cleaner--Universal Storage works with Any Platform.
Removed section on Deprecated storage types--if a developer is already using those, they can contact us directly. Showing the JSON storage type under Deprecated was particularly confusing, because we still allow JSON uploads to our other storage types (there's just a specific URL with JSON in it that we don't support anymore).